% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gibbs_chain_C}
\alias{Gibbs_chain_C}
\title{A Gibbs sampler function using Rcpp}
\arguments{
\item{a}{one of the parameters of binary density}

\item{b}{one of the parameters of binary density}

\item{n}{one of the parameters of binary density}

\item{N}{the number of samples generated}

\item{burn}{the number of samples discarded}

\item{x0}{the initial x-value}

\item{y0}{the initial y-value}
}
\value{
A NumericMatrix of the samples \code{n}
}
\description{
Samples are drawn according to the idea of Gibbs sampler using Rcpp
}
\examples{
\dontrun{
    set.seed(12345)
    N <- 1e4
    burn <- 1000
    a <- 2
    b <- 4
    n <- 16
    x0 <- sample(0:n,1)
    y0 <- runif(1)
    xC <- Gibbs_chain_C(a, b, n, N, burn, x0, y0)
}
}
