% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_aco.R
\name{My_aco}
\alias{My_aco}
\title{Ant Colony Optimization using R}
\usage{
My_aco(cities, AntsNum, max_iter = 100, alpha, beta, rho, Q)
}
\arguments{
\item{cities}{the coordinates of the cities}

\item{AntsNum}{the number of the ants}

\item{max_iter}{the maximum number of iterations for optimizing, the default value is 100.}

\item{alpha}{the pheromone factor}

\item{beta}{the heuristic factor}

\item{rho}{the pheromone volatile factor}

\item{Q}{the pheromone increment}
}
\value{
a list of the optimal path in each iteration and its length and the best path with its length \code{n}
}
\description{
Calculate the best path with ACO algorithm and show the iterative process by a figure using R
}
\examples{
\dontrun{
    set.seed(123)
    n <- 20
    alpha <- 1
    beta <- 2
    rho <- 0.5
    Q <- 100
    AntsNum <- 10
    max_iter <- 100
    cities <- data.frame(x = runif(n, min = 0, max = 100),
                         y = runif(n, min = 0, max = 100))
    result <- My_aco(cities, AntsNum, max_iter, alpha, beta, rho, Q)
    print(result$min_bestPath)
    print(result$min_bestLength)
    plot(
      result$bestLength,
      type = "l",
      main = "蚁群算法用例",
      xlab = "迭代次数",
      ylab = "最佳路径长度"
    )
}
}
