% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_kmeans_plusplus.R
\name{My_kmeans_plusplus}
\alias{My_kmeans_plusplus}
\title{Calculate the center of K-mean++ cluster using R}
\usage{
My_kmeans_plusplus(data, k, max_iter = 100)
}
\arguments{
\item{data}{the data for clustering}

\item{k}{the number of the centers}

\item{max_iter}{the maximum number of iterations for clustering, the default value is 100.}
}
\value{
a list of the centers and clusters \code{n}
}
\description{
Calculate the centers of K-mean++ cluster and show the clustering results by a figure using R
}
\examples{
\dontrun{
    set.seed(123)
    data <- matrix(rnorm(100 * 2), nrow = 100)
    k <- 3
    result <- My_kmeans_plusplus(data, k)
    print(result$centers)
    print(result$clusters)
}
}
