\name{SA24168406-package}
\alias{SA24168406-package}
\docType{package}
\title{This package provides 4 functions for final project }
\description{R and Rcpp versions of kmeans++, aco, stratified sampler and Gibbs sampler function are }
\author{
Yang Chengming
Maintainer: Yang Chengming <15052271101@163.com>
}
\keyword{ package }
